Exploration of full dataset (redownload)
After a redownload of the Flickr Export data, I have much more content.
So it is more likely that it is complete. Needed to perform some data
analysis first to see whether there are additional special cases I
need to take care of.

In order to do, refactored a bit to not use the command line runner
anymore, but trigger steps via JUnit. Makes it easier to execute various
'commands'.

Performed an analysis to check whether photos occur in multiple Albums, as
I am considering switching to a move i.o. copy for storage and performance
reasons. But there are double uses, so will require a different strategy
to make that work.

Also checked whether I now have matching content with metadata. In the
process discovered there were actually 2 filename formats for content. In
which the IDs are positioned elsewhere. The order of the patterns is even
important in order to attribute it correctly. After this was fixed, there
were no more mismatches so all content items have a metadata item !