Investigation of the Uncategorized
Investigation of the Uncategorized required some updates to the name creation.
Temporarily introduced and removed the photo id as it helps for debugging the Flickr dataset, but
is not very friendly once the album structure is created.

As part of that investigation discovered there were some 'technical' tags associated with some photos.
Since they seemed to add little, I have 'ignored' them.
In addition chose for a different encoding of the tags, that should make it easier to detect them as
being tags and better retains the separation between individual tags.

The Uncategorized cases seem to stem from inconsistencies in the albums.json as for example Liesbeths vierde
maand is indicated to have only 45 photos, while when I go to Flickr online and look at my own photo numbering
we have several more... Which then end up 'Uncategorized'. Since the dataset is incorrect and not really my
parsing, will just keep it as is and do a manual correction myself afterwards.