Considerable refactoring towards meta/content separation
In order to be robust under retries while moving files, I had to ensure I base most my work on the metadata
and not on the content index. This required quite some rework. This improved separation also allowed to
cleanup and extract some things.

Extracted the FileNameSanitizer to hold the shared logic to cleanup a String for inclusion in a filename.
Extracted Move and Copy functionality (which follow the same pattern) into a Transfer enum.
Applied a different computation for the filename to take into account name overlaps.
Allowed the StructuringService to no longer refer to other services. This coordination is now external.

Discovered that basing the file names entirely on metadata is not a great idea for a subset of the photos.
Will have to devise a fallback strategy...